findIntegrity/requirements.txt
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-localflavor==2.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
importlib-metadata==1.5.0
Markdown==3.2.1
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.5
pytest-flake8==1.0.4
python-decouple==3.3
python-stdnum==1.13
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
wcwidth==0.1.8
zipp==2.2.0
